requirements.txt
streamlitstreamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
mlflow
shap
xgboost
ydata-profiling
pycaret